Refactored fully these sections into smaller components. Fixed cheetoh positioning for bottom line.
