+++ src/components/AboutUs/TheBottomLineSection.js
@@ -1,7 +1,15 @@
 import React from "react";
+import Hidden from '@material-ui/core/Hidden';
+import Grid from '@material-ui/core/Grid';
 
+import GridContainer from '../shared/GridContainer'
+import TextBox from '../shared/TextBox'
+import Title from '../shared/Title'
+import PContent from '../shared/ParagraphedContent'
+import Question from '../shared/Question'
+import ContactButton from '../shared/ContactButton'
 import People from '../../images/about-us/bottom-line/people.svg';
-import GridContainerTextPic from '../shared/GridContainerTextPic';
+import SVGLoad from '../shared/SVG'
 
 export default function TheBottomLineSection() {
 
@@ -11,7 +19,25 @@ export default function TheBottomLineSection() {
     const buttonText = "Contact Us"
 
     return (
-        <GridContainerTextPic title={title} content={content} displayCheetoh={true} displayHiddenComponent={true} hiddenQuestion={question} hiddenButtonText={buttonText} image={People}/>
+        <GridContainer>
+            <Grid item xs={12} md={6}>
+            <TextBox>
+                <Title title={title} hideCheetohs={["lg","md"]}/>
+                <PContent content={content} hideCheetohs={["sm","xs"]}/>
+                <Hidden smDown>
+                    <Question question={question}/>
+                    <ContactButton buttonText={buttonText}/>
+                </Hidden>
+            </TextBox>
+            </Grid>
+            <Grid item xs={12} md={6}>
+                <SVGLoad image={People} alt="peopleSVG" />
+                <Hidden mdUp>
+                    <Question question={question}/>
+                    <ContactButton buttonText={buttonText}/>
+                </Hidden>
+            </Grid>
+        </GridContainer>
     )
 };
