+++ src/components/shared/Question.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { useStyles } from './styles/QuestionStyle'
+
+export default function QuestionText(props){
+    const classes = useStyles();
+
+    return (
+        <p className={classes.question}>{props.question}</p>
+    )
+}
