+++ src/components/shared/TextBox.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { useStyles } from './styles/TextBoxStyle'
+
+export default function TextBox(props){
+    const classes = useStyles();
+
+    return(
+        <div className={classes.textBox}>
+            {props.children}
+        </div>
+    )
+}
