+++ src/components/shared/styles/GridContainerStyle.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+    root: {
+        minHeight: '500px',
+        [theme.breakpoints.down('sm')]: {
+            textAlign: 'center',
+        }
+    }
+}));
+
+export { useStyles };
