+++ src/components/shared/styles/SVGStyle.js
@@ -1,0 +1,13 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+    svgSettings: {
+        width: '80%',
+        [theme.breakpoints.up('sm')]: {
+            padding: '10%',
+            // width: '100%'
+        },
+    }
+}));
+
+export { useStyles };
