+++ src/components/shared/styles/TextBoxStyle.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+    textBox: {
+        padding: '0px 10%',
+        [theme.breakpoints.down('sm')]: {
+            paddingTop: '50px',
+        }
+    }
+}));
+
+export { useStyles };
